pyproject.toml
[project]
name = "file-transfer-automation"
authors = [{ name = "Fredrik Haglund", email = "fr3h4h@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10.0"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version"]

dependencies = [
    "smbprotocol==1.10.1",
    "fastapi==0.95.0",
    "uvicorn==0.21.1",
    "PyMySQL==1.0.2",
    "python_dotenv==1.0.0",
    "scheduleplus",
]

[project.optional-dependencies]
test = ["ruff"]

[tool.ruff]
target-version = "py310"

select = [
    "B007", # Loop control variable {name} not used within loop body
    "B014", # Exception handler with duplicate exception
    # "C",    # complexity
    "D",       # docstrings
    "E",       # pycodestyle
    "F",       # pyflakes/autoflake
    "ICN001",  # import concentions; {name} should be imported as {asname}
    "PGH004",  # Use specific rule codes when using noqa
    "PLC0414", # Useless import alias. Import alias does not rename original package.
    "SIM105",  # Use contextlib.suppress({exception}) instead of try-except-pass
    "SIM117",  # Merge with-statements that use the same scope
    "SIM118",  # Use {key} in {dict} instead of {key} in {dict}.keys()
    "SIM201",  # Use {left} != {right} instead of not {left} == {right}
    "SIM212",  # Use {a} if {a} else {b} instead of {b} if not {a} else {a}
    "SIM300",  # Yoda conditions. Use 'age == 42' instead of '42 == age'.
    "SIM401",  # Use get from dict with default instead of an if block
    "T20",     # flake8-print
    "TRY004",  # Prefer TypeError exception for invalid type
    "RUF006",  # Store a reference to the return value of asyncio.create_task
    "UP",      # pyupgrade
    "W",       # pycodestyle
]

ignore = [
    "D202", # No blank lines allowed after function docstring
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "D406", # Section name should end with a newline
    "D407", # Section name underlining
    "E501", # line too long
    "E731", # do not assign a lambda expression, use a def
    # Ignored due to performance: https://github.com/charliermarsh/ruff/issues/2923
    "UP038", # Use `X | Y` in `isinstance` call instead of `(X, Y)`
]
